requirements.txt
asgiref==3.7.2
autopep8==2.0.4
Django==5.0.2
django-ckeditor==6.7.1
django-js-asset==2.2.0
django-silk==5.1.0
gprof2dot==2022.7.29
pillow==10.2.0
pycodestyle==2.11.1
pyspellchecker==0.8.1
sqlparse==0.4.4
tzdata==2024.1
